Improve performance of CachedClipboard::findMatchingAbility
Pluck is a relatively expensive operation and in this case all the functionality it provides, is not needed (for example dot notation).

The combination of toBase() & pluck() is wasteful, and instead we can simply map the collection, use the ID as the key and called the identifier attribute to get the result we need.

This increased performance by around 40% during testing.